Соединил контроллеры в один класс
